requirements.txt
dagster==1.10.17
dagster-webserver==1.10.17
dagster-graphql==1.10.17
dagster_postgres==0.26.17
mariadb
psycopg
psycopg_pool
psycopg2-binary
pandas==2.1.3
sqlalchemy==2.0.23
slack-sdk
pandas_gbq
python-dotenv==1.0.0
pyodbc
statsmodels==0.14.0
oauth2client==4.1.3
google-api-python-client==2.153.0
prophet==1.1.5
numba==0.58.1
gspread==5.12.0
xmltodict==0.13.0
tableau_api_lib
sql_metadata==2.10.0
urllib3
playwright==1.39.0
beautifulsoup4==4.12.2
Authlib==1.2.1
pycryptodome==3.20.0
itunes-app-scraper-dmi
google-play-scraper
translators==5.8.9
retrying==1.3.4
python-slack-logger
requests==2.31.0
clickhouse_driver
clickhouse_connect
httplib2==0.22.0
certifi==2023.11.17
elasticsearch==8.12.0
pytz==2024.1
sqlparse==0.4.4
python-gitlab
pyarrow
pytrends==4.9.2
prometheus-api-client==0.5.5
serpapi==0.1.5
google-search-results==2.4.2
hvac==2.3.0
deep-translator==1.11.4
dask
dask[dataframe]
dask[distributed]
tableauserverclient
bokeh!=3.0.*,>=2.4.2
google-ads==25.1.0
trino
google-auth-httplib2
google-auth-oauthlib
simple-salesforce==1.12.6
mysql-connector-python== 9.1.0
fireducks==1.2.6